BREAKING CHANGE: update types to latest version `reefjs@12`
