+++ types/reefjs/reefjs-tests.ts
@@ -1,113 +1,42 @@
-import Reef = require("reefjs");
+import { store, render, component } from 'reefjs';
 
-// Turns debugging mode on
-Reef.debug(true);
+// Component re-render interactions with multiple store instances
+(() => {
+    const reefStore = store({ text: 'Hello' }); // fires the default reef:store event on change
+    const fooStore = store({ text: 'World' }, 'foo'); // fires reef:store-foo on change
+    const barStore = store({ text: '!' }, 'bar'); // fires reef:store-bar on change
 
-// Hello World
-const app = new Reef('#someelement', {
-    data: {
-        sayHello: 'Hello world',
-    },
+    const templateFn = () => `<p>${reefStore.text}, ${fooStore.text}${barStore.text}</p>`;
 
-    template: props => {
-        return `<h1>${props.sayHello} from Reef!</h1>`;
-    },
-});
+    component('#app', templateFn, { stores: ['foo', 'bar'], events: false });
+    // listens only to reef:store-foo and reef:store-bar
+    // initially renders Hello, World!
 
-app.render();
+    setTimeout(() => {
+        fooStore.text = 'Earth'; // fires reef:store-foo
+    }, 1000); // component re-renders, renders Hello, Earth!
 
-// Nested components
-const nc = new Reef('#nc', {
-    data: {
-        id: 'test',
-    },
+    setTimeout(() => {
+        reefStore.text = 'Hi'; // fires reef:store
+    }, 2000); // component does not re-render, DOM still shows Hello, Earth!
 
-    template: props => {
-        return `
-        <div id="${props.id}">
-            <p id="name1"></p>
-            <p id="name2"></p>
-        </div>
-        `;
-    },
-});
+    setTimeout(() => {
+        barStore.text = '?'; // fires reef:store-bar
+    }, 3000); // component re-renders, renders Hi, Globe?
+})();
 
-const nc2 = new Reef('#name1', {
-    template: () => {
-        return 'My name is David!';
-    },
-});
+// Safely render UI components test
+(() => {
+    const elem = document.getElementById('test4') as HTMLElement;
+    const templateIframe = `<iframe src="data:text/html,<p>test 1, you should not see this</p>"></iframe>`;
+    const templateAnchor = `<a href="javascript:(function(){ document.getElementById('test2').innerHTML = 'ohno! this should not happen' })();">test2</a>`;
+    const templateButton = `<button onclick="(function(el){ el.innerHTML = 'inlined events are enabled!' })(this)">click me</button>`;
 
-const nc3 = new Reef('#name2', {
-    data: {
-        name: 'Jane',
-    },
+    render('#test1', templateIframe, undefined); // should render an empty iframe
 
-    template: props => {
-        return `My name is ${props.name}!`;
-    },
-});
+    render('#test2', templateAnchor, true); // should still render an empty anchor tag
 
-nc.attach([nc2, nc3]);
-nc.render();
+    render('#test3', templateButton); // should render a button with no onclick listener
 
-// Edit data
-nc.data.name = 'Daniel!';
-
-// Detach component
-nc.detach(nc3);
-
-// Clone Data
-const copyOfData = Reef.clone(app.data);
-copyOfData.sayHello = 'Welcome message';
-
-const app2 = new Reef('#someelement2', {
-    data: {
-        sayHello: 'Hello world',
-    },
-
-    template: (props) => {
-        return `<h1>${props.sayHello} from Reef!</h1>`;
-    },
-});
-
-app2.render();
-
-app2.data = copyOfData;
-
-// Shared data
-const store = new Reef.Store({
-    data: {
-        todos: ['Swim', 'Climb', 'Jump', 'Play'],
-    },
-
-    setters: {
-        addTodo: (props, todo) => {
-            props.todos.push(todo);
-        },
-    },
-
-    getters: {
-        total: (props) => {
-            return props.todos.length;
-        },
-    },
-});
-
-const app3 = new Reef('#app', {
-    store,
-    template: (props) => {
-        return `<h1>${props.heading}</h1>
-        <ul>
-			${props.todos
-                .map((todo: string) => {
-                    return `<li>${todo}</li>`;
-                }).join('')}
-		</ul>`;
-    },
-});
-
-app3.render();
-
-store.do('addTodo', 'Sleep');
-store.get('total');
+    render(elem, templateButton, true); // should render an button that replaces DOM with `inlined events are enabled!` on click
+})();
