+++ types/reefjs/tsconfig.json
@@ -1,14 +1,11 @@
 {
     "compilerOptions": {
+        "strict": true,
         "module": "commonjs",
         "lib": [
-            "es6", 
+            "es6",
             "dom"
         ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictFunctionTypes": true,
-        "strictNullChecks": true,
         "baseUrl": "../",
         "typeRoots": [
             "../"
